Add bootfile_name (67) text as vendor specific option
while running nios_network in check mode, nios_network will find a use_option parameter in the provided bootfile_name option, and no use_option on the existing bootfile_name.
Suggesting to add 67 into the check_vendor_specific_dhcp_option hack, that will allow to remove the use_option key on the bootfile_name option, providing it is the first one of the list of options.